refactored ownCommunities
